Add support for Safari
As part of supporting >1 browser, add a compatibility layer for Safari/Chrome.

Safari and Chrome have different APIs to run Javascript in a specific tab.

Therefore, we now wrap the "raw" tab found via Applescript in a custom JS facade
object to provide a standard API. The facade knows which application it's for
(Safari or Chrome) and will run provided JS accordingly.